power: charge animation: Support assign regulator mem state
Vdd_logic is default disabled when kernel system suspend
on some platform, and regulator suspend off is configured
by default in DTS. But U-Boot can't support that deep system
suspend. So prividing the prop to assign what we really
want regulators to be in U-Boot system suspend.

Example:
    regulator-on-in-mem = <&vcc_ddr>, <&vdda0v9_pmu>,
			  <&vcca1v8_pmu>, <&vcc3v3_pmu>;
    regulator-off-in-mem = <&vdd_logic>, <&vdd_gpu>,
			   <&vcc_ddr>, <&vcc_3v3>, <&vdda_0v9>;

If not finding these prop, fall through to the DTS configure
in every regulator node.

Signed-off-by: Joseph Chen <chenjh@rock-chips.com>
Change-Id: Ie40c232317a98e7621ad1704b14718415f73a35a